upgrade zemu test to ts (wip)
