--- tests_zemu/jest.js
@@ -1,2 +1,0 @@
-export default jest;
-export const { expect, test } = global;
