+++ tests_zemu/package.json
@@ -4,39 +4,43 @@
   "license": "Apache-2.0",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "jest --detectOpenHandles -t 'Standard'",
-    "testval": "jest --detectOpenHandles -t 'Validator'"
-  },
+  "types": "./dist/index.d.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Zondax/ledger-oasis"
   },
   "keywords": [
-    "zondax"
+    "Zondax",
+    "Ledger"
   ],
+  "scripts": {
+    "test": "jest --detectOpenHandles -t 'Standard'",
+    "testval": "jest --detectOpenHandles -t 'Validator'"
+  },
   "dependencies": {
     "@zondax/ledger-oasis": "file:../js",
-    "@zondax/zemu": "^0.12.3"
+    "@zondax/zemu": "^0.13.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.13.16",
-    "@babel/core": "^7.13.16",
-    "@babel/eslint-parser": "^7.13.14",
-    "@babel/node": "^7.13.13",
-    "@babel/plugin-transform-runtime": "^7.13.15",
-    "@babel/preset-env": "^7.13.15",
-    "babel-jest": "26.6.3",
+    "@types/jest": "^26.0.23",
+    "@types/ledgerhq__hw-transport": "^4.21.3",
+    "@types/secp256k1": "^4.0.2",
+    "@typescript-eslint/eslint-plugin": "^4.20.0",
+    "@typescript-eslint/parser": "^4.20.0",
+    "blakejs": "^1.1.0",
     "crypto-js": "4.0.0",
-    "ed25519-supercop": "^2.0.1",
-    "eslint": "^7.24.0",
-    "eslint-config-prettier": "^8.2.0",
+    "eslint": "^7.8.1",
+    "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.3.5",
-    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-jest": "^24.3.6",
+    "eslint-plugin-prettier": "^3.3.1",
     "jest": "26.6.3",
     "jest-serial-runner": "^1.1.0",
+    "secp256k1": "^4.0.2",
+    "prettier": "^2.0.4",
+    "ts-jest": "^26.5.4",
+    "typescript": "^4.2.3",
+    "ed25519-supercop": "^2.0.1",
     "js-sha512": "^0.8.0"
   }
 }
