+++ tests_zemu/snapshots/s-mainmenu/00000.png
